srealityscraper/requirements.txt
cryptography==3.3.2
scrapy==2.6.2
scrapy-splash==0.8.0
psycopg2-binary==2.9.3
